+++ paramdb/__init__.py
@@ -2,18 +2,20 @@
 Database for storing and retrieving QPU parameters during quantum control experiments.
 """
 
+from paramdb._param_data._param_data import ParamData
+from paramdb._param_data._dataclasses import Param, Struct
+from paramdb._param_data._collections import ParamList, ParamDict
+from paramdb._param_data._type_mixins import ParentType, RootType
 from paramdb._database import ParamDB, CommitEntry
-from paramdb._param_data import ParamData, Struct, Param
-from paramdb._exceptions import CommitNotFoundError
-
-__version__ = "0.1.0"
 
 __all__ = [
-    "__version__",
+    "ParamData",
+    "Param",
+    "Struct",
+    "ParamList",
+    "ParamDict",
+    "ParentType",
+    "RootType",
     "ParamDB",
     "CommitEntry",
-    "ParamData",
-    "Struct",
-    "Param",
-    "CommitNotFoundError",
 ]