+++ tests/helpers.py
@@ -1,8 +1,47 @@
 """Helper functions for paramdb tests."""
 
+from __future__ import annotations
+from typing import Any
+from dataclasses import dataclass, field
 import time
-from datetime import datetime
-from tests.conftest import CustomStruct, CustomParam
+from paramdb import ParamData, Param, Struct, ParamList, ParamDict
+
+DEFAULT_NUMBER = 1.23
+DEFAULT_STRING = "test"
+
+
+@dataclass
+class CustomParam(Param):
+    """Custom parameter."""
+
+    number: float = DEFAULT_NUMBER
+    number_init_false: float = field(init=False, default=DEFAULT_NUMBER)
+    string: str = DEFAULT_STRING
+
+
+@dataclass
+# pylint: disable-next=too-many-instance-attributes
+class CustomStruct(Struct):
+    """Custom parameter structure."""
+
+    number: float = DEFAULT_NUMBER
+    number_init_false: float = field(init=False, default=DEFAULT_NUMBER)
+    string: str = DEFAULT_STRING
+    list: list[Any] = field(default_factory=list)
+    dict: dict[str, Any] = field(default_factory=dict)
+    param: CustomParam | None = None
+    struct: CustomStruct | None = None
+    param_list: ParamList[Any] = field(default_factory=ParamList)
+    param_dict: ParamDict[Any] = field(default_factory=ParamDict)
+    param_data: ParamData | None = None
+
+
+class CustomParamList(ParamList[Any]):
+    """Custom parameter list subclass."""
+
+
+class CustomParamDict(ParamDict[Any]):
+    """Custom parameter dictionary subclass."""
 
 
 def sleep_for_datetime() -> None:
@@ -15,20 +54,4 @@ def sleep_for_datetime() -> None:
     computers execute instructions faster than this. So without waiting, it is difficult
     to ensure that something is using `datetime.now()`.
     """
-    time.sleep(0.001)  # Wait for one millisecond
-
-
-def update_param_and_assert_last_updated_changed(
-    param: CustomParam, param_data: CustomStruct | CustomParam
-) -> None:
-    """
-    Update the given parameter (assumed to be or exist within the given parameter data)
-    and assert that the structure's last updated time correctly reflects that something
-    was just updated.
-    """
-    start = datetime.now()
-    sleep_for_datetime()
-    param.number += 1
-    sleep_for_datetime()
-    end = datetime.now()
-    assert param_data.last_updated is not None and start < param_data.last_updated < end
+    time.sleep(100e-6)  # Wait for 100 microseconds